Fixed utils pipeline test
